feat: 각 endpoint 인가 처리, etc.
각 endpoint에 맞는 인가 처리를 해주었다.
인증된 정보를 받아서 처리해야 하는 auditing 기능과 history API를 구현하였다.
문의, 후기 응답에 작성자 이름을 같이 보내주도록 구현하였다.
주소를 하나의 값으로 합치고 우편번호, 연락처는 제거하였다.

+++ server/src/main/java/com/codestates/seb41_main_034/question/QuestionFacade.java
@@ -1,6 +1,8 @@
 package com.codestates.seb41_main_034.question;
 
 import com.codestates.seb41_main_034.common.JsonListHelper;
+import com.codestates.seb41_main_034.common.exception.BusinessLogicException;
+import com.codestates.seb41_main_034.common.exception.ExceptionCode;
 import com.codestates.seb41_main_034.common.response.PaginatedData;
 import com.codestates.seb41_main_034.product.ProductService;
 import com.codestates.seb41_main_034.product.entity.Product;
@@ -8,13 +10,18 @@ import com.codestates.seb41_main_034.question.dto.AnswerRequestDto;
 import com.codestates.seb41_main_034.question.dto.QuestionDto;
 import com.codestates.seb41_main_034.question.dto.QuestionPatchDto;
 import com.codestates.seb41_main_034.question.dto.QuestionPostDto;
+import com.codestates.seb41_main_034.question.entity.Answer;
 import com.codestates.seb41_main_034.question.entity.Question;
+import com.codestates.seb41_main_034.user.entity.User;
+import com.codestates.seb41_main_034.user.service.UserService;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
+import java.util.HashSet;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
@@ -29,7 +36,7 @@ public class QuestionFacade {
 
     private final ProductService productService;
 
-    // TODO: 작성자 이름을 불러오기 위해 UserService 주입 필요
+    private final UserService userService;
 
     private final JsonListHelper helper;
 
@@ -40,19 +47,28 @@ public class QuestionFacade {
         // 엔티티 객체 생성
         Question question = questionService.createQuestion(questionPostDto);
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
         // DTO 매핑 후 반환
-        return question.toDto(helper, product);
+        return question.toDto(helper, product, Map.of(user.getId(), user.getMaskedName()));
     }
 
     public QuestionDto readQuestion(long questionId) {
         // 문의 조회
         Question question = questionService.readQuestion(questionId);
 
+        Set<Integer> userIds = new HashSet<>();
+        userIds.add(question.getCreatedBy());
+        Optional.ofNullable(question.getAnswer()).map(Answer::getCreatedBy).ifPresent(userIds::add);
+
         // 상품 정보 조회
         Product product = productService.readProduct(question.getProductId());
 
+        Map<Integer, String> idNameMap = userService.getVerifiedUsers(userIds).stream()
+                .collect(Collectors.toMap(User::getId, User::getMaskedName));
+
         // DTO 매핑 후 반환
-        return question.toDto(helper, product);
+        return question.toDto(helper, product, idNameMap);
     }
 
     public PaginatedData<QuestionDto> readProductQuestions(int productId, Pageable pageable) {
@@ -61,13 +77,22 @@ public class QuestionFacade {
 
         // 문의 목록 Page 조회
         Page<Question> questionPage = questionService.readProductQuestions(productId, pageable);
+        Set<Integer> userIds = new HashSet<>();
+        questionPage.getContent().forEach(question -> {
+            userIds.add(question.getCreatedBy());
+            Optional.ofNullable(question.getAnswer()).map(Answer::getCreatedBy).ifPresent(userIds::add);
+        });
+
+        Map<Integer, String> idNameMap = userService.getVerifiedUsers(userIds).stream()
+                .collect(Collectors.toMap(User::getId, User::getMaskedName));
 
         // DTO 매핑 후 반환
-        return PaginatedData.of(questionPage.map(question -> question.toDto(helper, product)));
+        return PaginatedData.of(questionPage.map(question -> question.toDto(helper, product, idNameMap)));
     }
 
-    public PaginatedData<QuestionDto> readQuestionHistory(
-            int createdBy, LocalDate from, LocalDate to, Pageable pageable) {
+    public PaginatedData<QuestionDto> readQuestionHistory(LocalDate from, LocalDate to, Pageable pageable) {
+        int createdBy = ((User) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getId();
+
         // from, to 기본값 설정
         from = Optional.ofNullable(from).orElse(LocalDate.of(2023, 1, 1));
         to = Optional.ofNullable(to).orElse(LocalDate.now());
@@ -80,20 +105,46 @@ public class QuestionFacade {
         Map<Integer, Product> productMap = productService.getVerifiedProducts(productIds).stream()
                 .collect(Collectors.toMap(Product::getId, Function.identity()));
 
+        Set<Integer> userIds = new HashSet<>();
+        questionPage.getContent().forEach(question -> {
+            userIds.add(question.getCreatedBy());
+            Optional.ofNullable(question.getAnswer()).map(Answer::getCreatedBy).ifPresent(userIds::add);
+        });
+
+        Map<Integer, String> idNameMap = userService.getVerifiedUsers(userIds).stream()
+                .collect(Collectors.toMap(User::getId, User::getMaskedName));
+
         // DTO 매핑 후 반환
         return PaginatedData.of(
-                questionPage.map(question -> question.toDto(helper, productMap.get(question.getProductId()))));
+                questionPage.map(question ->
+                        question.toDto(helper, productMap.get(question.getProductId()), idNameMap)));
     }
 
     public QuestionDto updateQuestion(long questionId, QuestionPatchDto questionPatchDto) {
+        Question question = questionService.readQuestion(questionId);
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!user.getRoles().contains("ADMIN") && user.getId() != question.getCreatedBy()) {
+            throw new BusinessLogicException(ExceptionCode.AUTH_FORBIDDEN);
+        }
+
         // 문의 수정 후 반환
-        Question question = questionService.updateQuestion(questionId, questionPatchDto);
+        Question updatedQuestion = questionService.updateQuestion(questionId, questionPatchDto);
 
         // DTO 매핑 후 반환
-        return question.toDto(helper);
+        return updatedQuestion.toDto(helper);
     }
 
     public void deleteQuestion(long questionId) {
+        Question question = questionService.readQuestion(questionId);
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!user.getRoles().contains("ADMIN") && user.getId() != question.getCreatedBy()) {
+            throw new BusinessLogicException(ExceptionCode.AUTH_FORBIDDEN);
+        }
+
         // 문의 삭제
         questionService.deleteQuestion(questionId);
     }
