+++ server/src/main/java/com/codestates/seb41_main_034/review/ReviewFacade.java
@@ -1,6 +1,8 @@
 package com.codestates.seb41_main_034.review;
 
 import com.codestates.seb41_main_034.common.JsonListHelper;
+import com.codestates.seb41_main_034.common.exception.BusinessLogicException;
+import com.codestates.seb41_main_034.common.exception.ExceptionCode;
 import com.codestates.seb41_main_034.common.response.PaginatedData;
 import com.codestates.seb41_main_034.common.storage.ImageStorageService;
 import com.codestates.seb41_main_034.product.ProductService;
@@ -8,9 +10,12 @@ import com.codestates.seb41_main_034.product.entity.Product;
 import com.codestates.seb41_main_034.review.dto.ReviewDto;
 import com.codestates.seb41_main_034.review.dto.ReviewPatchDto;
 import com.codestates.seb41_main_034.review.dto.ReviewPostDto;
+import com.codestates.seb41_main_034.user.entity.User;
+import com.codestates.seb41_main_034.user.service.UserService;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -31,7 +36,7 @@ public class ReviewFacade {
 
     private final ProductService productService;
 
-    // TODO: 작성자 이름을 불러오기 위해 UserService 주입 필요
+    private final UserService userService;
 
     private final ImageStorageService imageStorageService;
 
@@ -48,19 +53,23 @@ public class ReviewFacade {
         // 엔티티 객체 생성
         Review review = reviewService.createReview(reviewPostDto, imageUrls);
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
         // DTO 매핑 후 반환
-        return review.toDto(helper, product);
+        return review.toDto(helper, product, user.getMaskedName());
     }
 
-    public ReviewDto readReview(long questionId) {
+    public ReviewDto readReview(long reviewId) {
         // 후기 조회
-        Review review = reviewService.readReview(questionId);
+        Review review = reviewService.readReview(reviewId);
 
         // 상품 정보 조회
         Product product = productService.readProduct(review.getProductId());
 
+        User user = userService.findVerifiedUserById(review.getCreatedBy());
+
         // DTO 매핑 후 반환
-        return review.toDto(helper, product);
+        return review.toDto(helper, product, user.getMaskedName());
     }
 
     public PaginatedData<ReviewDto> readProductReviews(int productId, Pageable pageable) {
@@ -70,18 +79,25 @@ public class ReviewFacade {
         // 후기 목록 Page 조회
         Page<Review> reviewPage = reviewService.readProductReviews(productId, pageable);
 
+        Set<Integer> userIds = reviewPage.stream().map(Review::getCreatedBy).collect(Collectors.toSet());
+
+        Map<Integer, String> idNameMap = userService.getVerifiedUsers(userIds).stream()
+                .collect(Collectors.toMap(User::getId, User::getMaskedName));
+
         // DTO 매핑 후 반환
-        return PaginatedData.of(reviewPage.map(review -> review.toDto(helper, product)));
+        return PaginatedData.of(
+                reviewPage.map(review -> review.toDto(helper, product, idNameMap.get(review.getCreatedBy()))));
     }
 
-    public PaginatedData<ReviewDto> readReviewHistory(
-            int createdBy, LocalDate from, LocalDate to, Pageable pageable) {
+    public PaginatedData<ReviewDto> readReviewHistory(LocalDate from, LocalDate to, Pageable pageable) {
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
         // from, to 기본값 설정
         from = Optional.ofNullable(from).orElse(LocalDate.of(2023, 1, 1));
         to = Optional.ofNullable(to).orElse(LocalDate.now());
 
         // 회원의 후기 목록 Page 조회
-        Page<Review> reviewPage = reviewService.readReviewHistory(createdBy, from, to, pageable);
+        Page<Review> reviewPage = reviewService.readReviewHistory(user.getId(), from, to, pageable);
 
         // 후기 목록에 있는 상품 정보를 한 번에 조회
         Set<Integer> productIds = reviewPage.get().map(Review::getProductId).collect(Collectors.toSet());
@@ -89,11 +105,19 @@ public class ReviewFacade {
                 .collect(Collectors.toMap(Product::getId, Function.identity()));
 
         // DTO 매핑 후 반환
-        return PaginatedData.of(reviewPage.map(review -> review.toDto(helper, productMap.get(review.getProductId()))));
+        return PaginatedData.of(
+                reviewPage.map(review ->
+                        review.toDto(helper, productMap.get(review.getProductId()), user.getMaskedName())));
     }
 
     public ReviewDto updateReview(long reviewId, ReviewPatchDto reviewPatchDto, List<MultipartFile> images) {
         Review review = reviewService.readReview(reviewId);
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!user.getRoles().contains("ADMIN") && user.getId() != review.getCreatedBy()) {
+            throw new BusinessLogicException(ExceptionCode.AUTH_FORBIDDEN);
+        }
 
         // 이미지 삭제
         List<String> imageUrlList = helper.jsonToList(review.getImageUrls());
@@ -116,6 +140,12 @@ public class ReviewFacade {
     public void deleteReview(long reviewId) {
         // 후기 조회
         Review review = reviewService.readReview(reviewId);
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!user.getRoles().contains("ADMIN") && user.getId() != review.getCreatedBy()) {
+            throw new BusinessLogicException(ExceptionCode.AUTH_FORBIDDEN);
+        }
 
         // 이미지 삭제
         imageStorageService.delete(helper.jsonToList(review.getImageUrls()));
